Added terraform exmaples

+++ make-layer.sh
@@ -2,33 +2,22 @@
 
 # To run properly read README.md
 
-set -a
-source .env
-set +a
-
-# Parse arguments
-# while [[ "$#" -gt 0 ]]; do
-#     case $1 in
-#         --path)
-#             HOST_PATH="$2"
-#             shift 2
-#             ;;
-#         *)
-#             echo "Unknown parameter passed: $1"
-#             exit 1
-#             ;;
-#     esac
-# done
+if [ -z "$HOST_PATH" ] || [ -z "$ARTIFACT" ]; then
+  echo "Either HOST_PATH or ARTIFACT env variables is missing, using .env"
+  set -a
+  source .env
+  set +a
+fi
 
 if [ -z $HOST_PATH ]; then
-  echo "Path not provided, using ./layer"
+  echo "Path not provided in .env, using default value: layer"
   HOST_PATH=layer
 fi
 
 mkdir -p $HOST_PATH
 
 if [ -z $ARTIFACT ]; then
-  echo "Artifact name not provided, using opencv-python312-layer.zip"
+  echo "Artifact name not provided in .env, using default value: opencv-python312-layer.zip"
   ARTIFACT="opencv-python312-layer.zip"
 fi
 
@@ -38,6 +27,11 @@ fi
 if [ -e "$HOST_PATH/$ARTIFACT" ]; then
     echo "Layer already created, exiting"
     exit 0
+fi
+
+if [[ ! "$ARTIFACT" =~ \.zip$ ]]; then
+  echo "artifact_name should end with .zip, changing it for you.."
+  export ARTIFACT=$ARTIFACT.zip
 fi
 
 # Verify the path exists
@@ -64,8 +58,10 @@ fi
 CONTAINER_NAME="layer-builder-$(date +%s)"
 
 # Run Docker container with volume attached
-docker run -it --env-file .env --name "$CONTAINER_NAME" makelayer:latest
+docker run -it -e ARTIFACT="$ARTIFACT" --name "$CONTAINER_NAME" makelayer:latest
 
 docker cp "$CONTAINER_NAME:/lambda/layer/$ARTIFACT" "$HOST_PATH/$ARTIFACT"
 
-docker rm $CONTAINER_NAME
+if [ "$?" -eq 0 ]; then
+  docker rm $CONTAINER_NAME
+fi
